refactor(PI-9): Separacion de contextos de base de datos
