+++ InfrastructureTests/Database/ChatRepositoryTests.cs
@@ -1,6 +1,7 @@
 ﻿using Domain.Entities;
 using Domain.Exceptions;
 using Domain.ValueObjects;
+using Infrastructure.Database.Chats;
 using InfrastructureTests.Database.Tests;
 using Microsoft.EntityFrameworkCore;
 using System;
