+++ InfrastructureTests/Database/UnitOfWorkTests.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Infrastructure.Database.Chats;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
